+++ server/cli/root.go
@@ -11,7 +11,7 @@ import (
 	"go.uber.org/zap"
 )
 
-const version = "1.2.0"
+const version = "1.3.0"
 
 var rootCmd = &cobra.Command{
 	Use:     "go-url-shortener",